bigquery-search/pyproject.toml
[project]
name = "bigquery-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.31.0",
    "streamlit>=1.44.1",
]
